requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0

# Database & ORM
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication
Flask-Login==0.6.3

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# API Clients
spotipy==2.23.0
google-api-python-client==2.108.0
requests==2.31.0
